requirements.txt
aiohttp
beautifulsoup4
hachoir
numpy
ffmpeg-python
Pillow
git+git://github.com/pyrogram/pyrogram@master
requests
tgcrypto
git+https://github.com/yt-dlp/yt-dlp
olefile
filetype
tldextract
lk21
motor
https://github.com/Tinche/aiofiles/archive/master.zip
dnspython
